implemented train_sgd

+++ hw2/tests/hw2_tests.rs
@@ -229,7 +229,7 @@ fn test_cross_entropy_loss() {
         TensorData::from([[2.0f64, 1.0, 0.0], [0.0, 2.0, 1.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 2]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 2]), &DEVICE);
     let loss: f64 = cross_entropy_loss(y_pred, y).into_scalar();
     assert!((loss - 0.907_606).abs() < 1e-5);
 }
@@ -240,7 +240,7 @@ fn test_cross_entropy_loss_three_class_batch() {
         TensorData::from([[1.0f64, 0.0, -1.0], [2.0, 1.0, 0.0], [-1.0, 2.0, 1.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 2, 1]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 2, 1]), &DEVICE);
     let loss: f64 = cross_entropy_loss(y_pred, y).into_scalar();
     assert!((loss - 1.054741).abs() < 1e-4);
 }
@@ -253,7 +253,7 @@ fn test_error() {
         TensorData::from([[3.0f64, 1.0], [0.0, 2.0], [1.0, 1.0], [-1.0, 0.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 1, 1, 1]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 1, 1, 1]), &DEVICE);
     let err = error_rate(y_pred, y);
     assert!((err - 0.25).abs() < 1e-6);
 }
@@ -266,7 +266,7 @@ fn test_train_sgd_one_epoch() {
         TensorData::from([[2.0f64, 1.0], [1.0, 2.0], [-2.0, -1.0], [-1.0, -2.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i32, 1, 0, 0]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i64, 1, 0, 0]), &DEVICE);
 
     let w = train_sgd(x, y, 2, 1, 0.1, 2);
     assert_eq!(w.dims(), [2, 2]);
@@ -291,7 +291,7 @@ fn test_train_sgd() {
         TensorData::from([[2.0f64, 1.0], [1.0, 2.0], [-2.0, -1.0], [-1.0, -2.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i32, 1, 0, 0]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i64, 1, 0, 0]), &DEVICE);
 
     let w = train_sgd(x.clone(), y.clone(), 2, 20, 0.1, 2);
     assert_eq!(w.dims(), [2, 2]);
@@ -299,7 +299,7 @@ fn test_train_sgd() {
     // Verify predictions are correct after training
     let w_data: Vec<f64> = w.clone().into_data().to_vec().unwrap();
     let x_data: Vec<f64> = x.into_data().to_vec().unwrap();
-    let y_data: Vec<i32> = y.into_data().to_vec().unwrap();
+    let y_data: Vec<i64> = y.into_data().to_vec().unwrap();
     for i in 0..4 {
         let yi = y_data[i] as usize;
         let scores: Vec<f64> = (0..2)
@@ -322,14 +322,14 @@ fn test_train_sgd_fifteen_epochs() {
         TensorData::from([[2.0f64, 1.0], [1.0, 2.0], [-2.0, -1.0], [-1.0, -2.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i32, 1, 0, 0]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i64, 1, 0, 0]), &DEVICE);
 
     let w = train_sgd(x.clone(), y.clone(), 2, 15, 0.1, 2);
     assert_eq!(w.dims(), [2, 2]);
 
     let w_data: Vec<f64> = w.clone().into_data().to_vec().unwrap();
     let x_data: Vec<f64> = x.into_data().to_vec().unwrap();
-    let y_data: Vec<i32> = y.into_data().to_vec().unwrap();
+    let y_data: Vec<i64> = y.into_data().to_vec().unwrap();
     for i in 0..4 {
         let yi = y_data[i] as usize;
         let scores: Vec<f64> = (0..2)
@@ -353,7 +353,7 @@ fn test_train_sgd_fifteen_epochs() {
 #[test]
 fn test_cross_entropy_loss_single_sample() {
     let y_pred: Tensor<B, 2> = Tensor::from_data(TensorData::from([[1.0f64, 2.0, 3.0]]), &DEVICE);
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([2i32]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([2i64]), &DEVICE);
     let loss: f64 = cross_entropy_loss(y_pred, y).into_scalar();
     // -3.0 + log(exp(1)+exp(2)+exp(3)) = -3 + log(e + e^2 + e^3) ≈ -3 + 3.4076 = 0.4076
     assert!(loss.is_finite());
@@ -367,7 +367,7 @@ fn test_cross_entropy_loss_perfect_prediction() {
         TensorData::from([[100.0f64, 0.0, 0.0], [0.0, 100.0, 0.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 1]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 1]), &DEVICE);
     let loss: f64 = cross_entropy_loss(y_pred, y).into_scalar();
     assert!(loss.is_finite());
     assert!(
@@ -381,7 +381,7 @@ fn test_cross_entropy_loss_numerically_stable() {
     // Very large logits should not produce NaN or Inf
     let y_pred: Tensor<B, 2> =
         Tensor::from_data(TensorData::from([[1000.0f64, 1001.0, 999.0]]), &DEVICE);
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i32]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i64]), &DEVICE);
     let loss: f64 = cross_entropy_loss(y_pred, y).into_scalar();
     assert!(loss.is_finite(), "Loss must be finite for large logits");
     assert!(loss >= 0.0, "Cross-entropy loss must be non-negative");
@@ -393,7 +393,7 @@ fn test_error_rate_perfect() {
         TensorData::from([[10.0f64, 0.0], [0.0, 10.0], [0.0, 10.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 1, 1]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 1, 1]), &DEVICE);
     let err = error_rate(y_pred, y);
     assert!(
         (err - 0.0).abs() < 1e-6,
@@ -405,7 +405,7 @@ fn test_error_rate_perfect() {
 fn test_error_rate_all_wrong() {
     let y_pred: Tensor<B, 2> =
         Tensor::from_data(TensorData::from([[0.0f64, 10.0], [10.0, 0.0]]), &DEVICE);
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 1]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 1]), &DEVICE);
     let err = error_rate(y_pred, y);
     assert!(
         (err - 1.0).abs() < 1e-6,
@@ -416,7 +416,7 @@ fn test_error_rate_all_wrong() {
 #[test]
 fn test_train_sgd_zero_epochs() {
     let x: Tensor<B, 2> = Tensor::from_data(TensorData::from([[1.0f64, 2.0], [3.0, 4.0]]), &DEVICE);
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i32, 1]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([0i64, 1]), &DEVICE);
     let w = train_sgd(x, y, 2, 0, 0.1, 2);
     // With 0 epochs, weights should remain at initialization (zeros)
     let w_data: Vec<f64> = w.into_data().to_vec().unwrap();
@@ -434,7 +434,7 @@ fn test_train_sgd_batch_size_one() {
         TensorData::from([[2.0f64, 1.0], [1.0, 2.0], [-2.0, -1.0], [-1.0, -2.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i32, 1, 0, 0]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i64, 1, 0, 0]), &DEVICE);
     let w = train_sgd(x, y, 2, 10, 0.1, 1);
     assert_eq!(w.dims(), [2, 2]);
     let w_data: Vec<f64> = w.into_data().to_vec().unwrap();
@@ -450,7 +450,7 @@ fn test_train_sgd_loss_decreases() {
         TensorData::from([[2.0f64, 1.0], [1.0, 2.0], [-2.0, -1.0], [-1.0, -2.0]]),
         &DEVICE,
     );
-    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i32, 1, 0, 0]), &DEVICE);
+    let y: Tensor<B, 1, Int> = Tensor::from_data(TensorData::from([1i64, 1, 0, 0]), &DEVICE);
 
     let w1 = train_sgd(x.clone(), y.clone(), 2, 1, 0.1, 2);
     let w20 = train_sgd(x.clone(), y.clone(), 2, 20, 0.1, 2);
@@ -463,7 +463,7 @@ fn test_train_sgd_loss_decreases() {
     // Compute sum of correct predictions for w20 (should be better than w1)
     let mut correct_w1 = 0;
     let mut correct_w20 = 0;
-    let y_data: Vec<i32> = y.into_data().to_vec().unwrap();
+    let y_data: Vec<i64> = y.into_data().to_vec().unwrap();
     for i in 0..4 {
         let s1: Vec<f64> = (0..2)
             .map(|k| (0..2).map(|j| w1_data[k * 2 + j] * x_data[i * 2 + j]).sum())
